+++ plugin/src/main/resources/js/iframe/host/main.js
@@ -1,7 +1,7 @@
 /**
  * Entry point for xdm messages on the host product side.
  */
-_AP.define("host/main", ["_dollar", "_xdm", "host/_addons", "host/_status_helper", "messages/main"], function ($, XdmRpc, addons, statusHelper, messages) {
+_AP.define("host/main", ["_dollar", "_xdm", "host/_addons", "host/_status_helper", "messages/main", "_ui-params"], function ($, XdmRpc, addons, statusHelper, messages, uiParams) {
 
   var xhrProperties = ["status", "statusText", "responseText"],
       xhrHeaders = ["Content-Type"],
@@ -13,7 +13,30 @@ _AP.define("host/main", ["_dollar", "_xdm", "host/_addons", "host/_status_helper
     return $("#embedded-" + ns);
   }
 
+  /**
+  * @name Options
+  * @class
+  * @property {String}  ns            module key
+  * @property {String}  src           url of the iframe
+  * @property {String}  w             width of the iframe
+  * @property {String}  h             height of the iframe
+  * @property {String}  dlg           is a dialog (disables the resizer)
+  * @property {String}  simpleDlg     deprecated, looks to be set when a confluence macro editor is being rendered as a dialog
+  * @property {Boolean} general       is a page that can be resized
+  * @property {String}  productCtx    context to pass back to the server (project id, space id, etc)
+  * @property {String}  key           addon key from the descriptor
+  * @property {String}  uid           id of the current user
+  * @property {String}  ukey          user key
+  * @property {String}  data.timeZone timezone of the current user
+  * @property {String}  cp            context path
+  */
+
+  /**
+  * @param {Options} options These values come from the velocity template and can be overridden using uiParams
+  */
   function create(options) {
+
+    $.extend(options, uiParams.fromUrl(options.src));
 
     var ns = options.ns,
         homeId = "ap-" + ns,